Add Command Refactoring
