+++ src/main/java/seedu/duke/Duke.java
@@ -1,24 +1,62 @@
 package seedu.duke;
 
+import seedu.duke.command.Command;
 import seedu.duke.data.TransactionList;
+import seedu.duke.exception.MoolahException;
 
 import java.util.Scanner;
 
 public class Duke {
-    public static void main(String[] args) {
-        TransactionList transactions = new TransactionList();
-        String inData;
-        Scanner scan = new Scanner(System.in);
-        Ui.showGreeting();
+    private Storage storage;
+    private TransactionList transactions;
+    private Ui ui;
 
-        while (scan.hasNextLine() && true) {
-            // Receives user input continuously
-            inData = scan.nextLine();
-            inData = inData.trim();
+    public Duke() {    // NEED TO ADD FILE PATH
+        ui = new Ui();
+        transactions = new TransactionList();
 
-            if (!Parser.parse(inData, transactions)) {
-                break;
+        // ideal code after u add all the storage stuff
+        /**storage = new Storage(filePath);
+        try {
+            tasks = new TaskList(storage.load());
+        } catch (DukeException e) {
+            ui.showLoadingError();
+            tasks = new TaskList();
+        }**/
+    }
+
+    public void run() {
+        ui.showGreeting();
+        boolean isExit = false;
+        String inData; // temp
+        while (!isExit) {
+            try {
+                Scanner scan = new Scanner(System.in);
+                inData = scan.nextLine();
+                inData = inData.trim();
+
+                Command c = Parser.parse(inData);
+                c.execute(transactions, ui, storage);
+                isExit = c.isExit();
+            } catch (MoolahException e) {
+                Ui.showErrorMessage(e.getMessage());
             }
+            //ideal code
+            /***try {
+                String fullCommand = ui.readCommand();
+                ui.showLine(); // divider line
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e.getErrorMessage());
+            } finally {
+                ui.showLine();
+            }**/
         }
+    }
+
+    public static void main(String[] args) {
+        new Duke().run();  // NEED TO ADD FILE PATH
     }
 }
