+++ src/main/java/seedu/duke/Parser.java
@@ -1,284 +1,50 @@
 package seedu.duke;
 
-import seedu.duke.common.ErrorMessages;
-import seedu.duke.common.InfoMessages;
-import seedu.duke.data.TransactionList;
-import seedu.duke.exception.AddTransactionMissingTagException;
-import seedu.duke.exception.AddDeleteInvalidIndexException;
-import seedu.duke.exception.AddTransactionUnknownTypeException;
-import seedu.duke.exception.AddTransactionInvalidCategoryException;
-import seedu.duke.exception.AddTransactionInvalidDateException;
+import seedu.duke.command.*;
 import seedu.duke.exception.MoolahException;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.Scanner;
-
-import static seedu.duke.common.DateFormats.DATE_INPUT_PATTERN;
+import seedu.duke.exception.UnknownInputException;
 
 public class Parser {
-    static final boolean IS_EXIT = false;
-    static final boolean IS_CONTINUE = true;
 
     /**
      * Parses the user input and deal with any input error returned.
      *
      * @param inData       The user input.
-     * @param transactions The array which would be operated on.
-     * @return IS_EXIT If input equals "bye", else return IS_CONTINUE.
-     */
-    public static boolean parse(String inData, TransactionList transactions) {
-        boolean isContinue = IS_CONTINUE;
-        try {
-            isContinue = processInput(inData, transactions);
-        } catch (MoolahException exception) {
-            Ui.showErrorMessage(exception.getMessage());
-        }
-        return isContinue;
-    }
-
-    /**
-     * Parses the user input and processes it with the transactions arraylist.
-     *
-     * @param inData       The user input.
-     * @param transactions The array which would be operated on.
      * @return IS_EXIT If input equals "bye", else return IS_CONTINUE.
      * @throws MoolahException Any command input exceptions captured by Moolah Manager.
      */
-    private static boolean processInput(String inData, TransactionList transactions) throws MoolahException {
-        if (inData.equals("help")) {
-            Ui.showHelpList();
-        } else if (inData.equals("list")) {
-            // Prints all transactions if input is equal to "list"
-            String transactionsList = transactions.listTransactions();
-            if (transactionsList.isEmpty()) {
-                Ui.showInfoMessage(InfoMessages.MESSAGE_INFO_LIST_EMPTY.toString());
-                return IS_CONTINUE;
-            }
-            Ui.showTransactionsList(transactionsList, InfoMessages.MESSAGE_INFO_LIST.toString());
-        } else if (inData.equals("purge")) {
-            // Shows confirmation prompt before deleting all transactions
-            Ui.showInfoMessage(InfoMessages.MESSAGE_INFO_WARNING.toString());
-            Scanner confirmation = new Scanner(System.in);
-            String input = confirmation.nextLine();
-            if (input.equals("Y")) {
-                TransactionList.purgeEntries(transactions);
-            } else {
-                System.out.println(InfoMessages.MESSAGE_INFO_DIVIDER);
-                System.out.println("MOOOOOO.... Aborting Command, returning to Home.");
-                System.out.println(InfoMessages.MESSAGE_INFO_DIVIDER);
-            }
-        } else if (inData.equals(("bye"))) {
-            Ui.showExitMessage();
-            // Exits loop
-            return IS_EXIT;
-        } else if (inData.isBlank() || inData.isEmpty()) {
-            Ui.showInvalidCommand();
-        } else if (inData.contains(" ")) {
-            // Further parses the user input for user transaction commands
-            String[] userInput = inData.split(" ", 2);
-            String command = userInput[0];
-            switch (command) {
+    public static Command parse(String inData) throws MoolahException{
+        Command command = null;
+        String[] splitInput = inData.split(" ", 2);;
+
+        // list commands duke to list all the tasks stored and their completion status
+        // try at the start cos of the errors possibly
+        switch (splitInput[0]) {
+            case "help":
+                command = new HelpCommand();
+                break;
+            case "list":
+                command = new ListCommand();
+                break;
+            case "purge":
+                command = new PurgeCommand();
+                break;
             case "delete":
-                /*
-                Checks if userInput is in the correct input format by further parsing,
-                before adding entry to arraylist
-                */
-                parseDeleteIndex(userInput[1], transactions);
+                command = new DeleteCommand(splitInput[1]);
                 break;
             case "add":
-                /*
-                Checks if userInput is in the correct input format by further parsing,
-                before adding entry to arraylist
-                */
-                parseAddCommand(userInput[1], transactions);
+                command = new AddCommand(splitInput[1]);
                 break;
             case "edit":
-                /*
-                Checks if userInput is in the correct input format by further parsing,
-                before the entry in the arraylist
-                */
+                command = new EditCommand(splitInput[1]);
+                break;
+            case "bye":
+                command = new ByeCommand();
                 break;
             default:
-                // For invalid commands
-                Ui.showInvalidCommand();
-            }
-        } else {
-            // For any single-word inData , which are not valid commands
-            Ui.showInvalidCommand();
+                Ui.showInvalidCommand(); // if u still want this
+                throw new UnknownInputException();
         }
-        return IS_CONTINUE;
-    }
-
-    /**
-     * Parses the add transaction command by checking if the compulsory tags exist followed by adding the transaction.
-     * Then executes the command to add the transaction into the list.
-     *
-     * @param userInput The user input after the "add" command word.
-     * @throws AddTransactionMissingTagException Exceptions related to "add" command.
-     */
-    private static void parseAddCommand(String userInput, TransactionList transactions) throws MoolahException {
-        String[] splits = userInput.split(" ");
-        checkTagsExist(splits);
-        // TODO: To check that each parameter is in correct format, e.g. amount should be valid integer/double
-        // TODO: To move the add transaction logic below to Command class in Command.execute()
-        String description = "";
-        int amount = 0;
-        String category = "";
-        LocalDate date = null;
-        String type = "";
-        boolean inputIsValid = true;
-
-        for (String split : splits) {
-            String tag = split.substring(0, 2);
-            String parameter = split.substring(2);
-            switch (tag) {
-            case "t/":
-                type = parameter;
-                break;
-            case "c/":
-                try {
-                    parseCategoryTag(parameter);
-                    category = parameter;
-                } catch (AddTransactionInvalidCategoryException e) {
-                    Ui.printMessages(String.valueOf(ErrorMessages.MESSAGE_ERROR_ADD_COMMAND_INVALID_CATEGORY));
-                    inputIsValid = false;
-                }
-
-                break;
-            case "a/":
-                try {
-                    amount = Integer.parseInt(parameter);
-                } catch (NumberFormatException e) {
-                    Ui.showNonNumericError();
-                    inputIsValid = false;
-                }
-                break;
-            case "d/":
-                date = parseDateTag(parameter);
-                break;
-            case "i/":
-                description = parameter;
-                break;
-            default:
-                break;
-            }
-        }
-        if (inputIsValid) {
-            switch (type) {
-            case "expense":
-                transactions.addExpense(description, amount, category, date);
-                break;
-            case "income":
-                transactions.addIncome(description, amount, category, date);
-                break;
-            default:
-                throw new AddTransactionUnknownTypeException();
-            }
-        }
-    }
-
-    /**
-     * Processes the parameter. If it is an invalid parameter an exception error would be thrown.
-     *
-     * @param parameter The user input after the user tag.
-     * @throws AddTransactionInvalidCategoryException Invalid category parameter exception.
-     */
-    private static void parseCategoryTag(String parameter) throws AddTransactionInvalidCategoryException {
-        if (isNumeric(parameter)) {
-            throw new AddTransactionInvalidCategoryException();
-        }
-
-    }
-
-    /**
-     * Checks if the parameter contains numeric characters.
-     *
-     * @param parameter The user input after the user tag.
-     * @return true if there are numeric characters within the parameter.
-     */
-    public static boolean isNumeric(String parameter) {
-        char[] characters = parameter.toCharArray();
-        for (char character : characters) {
-            if (Character.isDigit(character)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-
-    /**
-     * Parse the user parameter input for date into a LocalDate object and returns it.
-     *
-     * @param parameter The user input after the user tag.
-     * @return The LocalDate object parsed from user input given.
-     * @throws AddTransactionInvalidDateException Invalid date format exception.
-     */
-    private static LocalDate parseDateTag(String parameter) throws AddTransactionInvalidDateException {
-        try {
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern(DATE_INPUT_PATTERN.toString());
-            LocalDate date = LocalDate.parse(parameter, formatter);
-            return date;
-        } catch (DateTimeParseException exception) {
-            throw new AddTransactionInvalidDateException();
-        }
-    }
-
-    /**
-     * Checks if the targeted tags exists in the split user inputs.
-     *
-     * @param splits The user input after the command word, split into a list for every space found.
-     * @throws AddTransactionMissingTagException Missing tag exception.
-     */
-    private static void checkTagsExist(String[] splits) throws AddTransactionMissingTagException {
-        // TODO: To add the tags into Command class instead
-        String[] tags = new String[]{"t/", "c/", "a/", "d/", "i/"};
-        for (String tag : tags) {
-            boolean found = findMatchingTagFromInputs(tag, splits);
-            if (!found) {
-                throw new AddTransactionMissingTagException();
-            }
-        }
-    }
-
-    /**
-     * Returns a boolean value on whether a tag can be found among the split user inputs.
-     *
-     * @param tag    A specific tag used to locate the command parameter.
-     * @param splits The user input after the command word, split into a list for every space found.
-     * @return Whether the tag is found within the split inputs.
-     */
-    private static boolean findMatchingTagFromInputs(String tag, String[] splits) {
-        boolean found = false;
-        for (String split : splits) {
-            if (split.startsWith(tag)) {
-                found = true;
-                break;
-            }
-        }
-        return found;
-    }
-
-    private static void parseDeleteIndex(String parameter, TransactionList transactions) throws MoolahException {
-        boolean isInputValid = true;
-        int index;
-        int numberOfTransactions;
-        numberOfTransactions = transactions.size();
-        try {
-            index = Integer.parseInt(parameter);
-        } catch (NumberFormatException e) {
-            Ui.showNonNumericError();
-            return;
-        }
-        if ((index > numberOfTransactions) || (index <= 0)) {
-            isInputValid = false;
-        }
-        if (isInputValid) {
-            TransactionList.deleteEntry(transactions, index);
-        } else {
-            throw new AddDeleteInvalidIndexException();
-        }
+        return command;
     }
 }
