+++ src/main/java/seedu/duke/Storage.java
@@ -1,0 +1,13 @@
+package seedu.duke;
+
+import seedu.duke.data.TransactionList;
+
+public class Storage {
+
+    public Storage() {
+
+    }
+
+    public void writeToFile(TransactionList transactions) {
+    }
+}
