+++ src/main/java/seedu/duke/command/HelpCommand.java
@@ -1,0 +1,17 @@
+package seedu.duke.command;
+
+import seedu.duke.Storage;
+import seedu.duke.Ui;
+import seedu.duke.data.TransactionList;
+
+public class HelpCommand extends Command {
+    @Override
+    public void execute(TransactionList transactions, Ui ui, Storage storage) {
+        Ui.showHelpList();
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
