+++ src/main/java/seedu/duke/exception/UnknownInputException.java
@@ -1,0 +1,15 @@
+package seedu.duke.exception;
+
+import seedu.duke.Ui;
+
+public class UnknownInputException extends MoolahException {
+
+    // HELP ME CHANGE PLS
+
+    @Override
+    public String getMessage() {
+        Ui.showInvalidCommand();
+        return "OOPS!!! I'm sorry, but I don't know what that means :-( ";
+    }
+
+}
